Move the definition of BaseTile into types.
