+++ include/types/mod.hpp
@@ -3,6 +3,7 @@
 
 #pragma once
 
+#include "types/base_tile.hpp"
 #include "types/global.hpp"
 #include "types/global_tile_iterator.hpp"
 #include "types/layout.hpp"
